refactor: Rename account related classes and interfaces to be in accordance with new naming scheme
