+++ src/main/java/com/lunark/lunark/controller/LoginController.java
@@ -2,8 +2,7 @@ package com.lunark.lunark.controller;
 
 import com.lunark.lunark.dto.AccountLoginDto;
 import com.lunark.lunark.dto.AuthenticationToken;
-import com.lunark.lunark.service.AccountService;
-import jakarta.servlet.http.HttpServlet;
+import com.lunark.lunark.service.IAccountService;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,7 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class LoginController {
 
     @Autowired
-    AccountService accountService;
+    IAccountService accountService;
 
     @PostMapping(path = "", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<AuthenticationToken> login(@RequestBody AccountLoginDto accountLoginDto, HttpServletResponse response) {
